Added Support for Teleporters. Added Support for Coins.
Altered StaticObject so it supports Animations, The end of world and
Coins are of this type, they can be visible/invisible and collidable/not
collidable. This object can be extended to create many types of powerups
or triggers.

Added Coins in first 2 maps.
Added teleports in first map.

Coins dont effect gameplay yet, maybe give more suntime?

Added Sounds for teleports and coins.

Refactored some code.